+++ package.json
@@ -50,16 +50,12 @@
     "hot-shots": "^8.3.1",
     "html-entities": "^1.2.1",
     "imurmurhash": "^0.1.4",
-    "is-url": "^1.2.4",
     "js-cookie": "^2.2.1",
     "js-yaml": "^4.1.0",
-    "linkinator": "^2.13.6",
     "liquidjs": "^9.22.1",
     "lodash": "^4.17.21",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.4.0",
-    "mdast-util-from-markdown": "^0.8.4",
-    "mkdirp": "^1.0.4",
     "morgan": "^1.10.0",
     "next": "^10.2.0",
     "node-fetch": "^2.6.1",
@@ -80,7 +76,6 @@
     "remark-parse": "^7.0.2",
     "remark-rehype": "^5.0.0",
     "revalidator": "^0.3.1",
-    "rimraf": "^3.0.2",
     "rss-parser": "^3.12.0",
     "search-with-your-keyboard": "1.1.0",
     "semver": "^7.3.5",
@@ -134,6 +129,7 @@
     "http-status-code": "^2.1.0",
     "husky": "^6.0.0",
     "image-size": "^1.0.0",
+    "is-url": "^1.2.4",
     "japanese-characters": "^1.1.0",
     "javascript-stringify": "^2.1.0",
     "jest": "^26.6.3",
@@ -141,9 +137,12 @@
     "jest-github-actions-reporter": "^1.0.3",
     "jest-silent-reporter": "^0.5.0",
     "jest-slow-test-reporter": "^1.0.0",
+    "linkinator": "^2.13.6",
     "make-promises-safe": "^5.1.0",
+    "mdast-util-from-markdown": "^0.8.4",
     "mime": "^2.4.4",
     "mini-css-extract-plugin": "^1.6.0",
+    "mkdirp": "^1.0.4",
     "mock-express-response": "^0.3.0",
     "mockdate": "^3.0.5",
     "nock": "^13.0.11",
@@ -155,6 +154,7 @@
     "process": "^0.11.10",
     "replace": "^1.2.1",
     "resolve-url-loader": "^4.0.0",
+    "rimraf": "^3.0.2",
     "robots-parser": "^2.3.0",
     "sass": "^1.32.13",
     "sass-loader": "^11.1.1",
