+++ script/helpers/add-redirect-to-frontmatter.js
