+++ script/i18n/homogenize-frontmatter.js
@@ -7,7 +7,7 @@
 //
 // [end-readme]
 
-const fs = require('fs/promises')
+const fs = require('fs').promises
 const path = require('path')
 const matter = require('gray-matter')
 const walk = require('walk-sync')
