+++ script/search/algolia-client.js
