+++ script/search/algolia-search-index.js
