+++ script/search/build-records.js
@@ -2,7 +2,7 @@ const domwaiter = require('domwaiter')
 const eventToPromise = require('event-to-promise')
 const chalk = require('chalk')
 const parsePageSectionsIntoRecords = require('./parse-page-sections-into-records')
-const languages = require('../languages')
+const languages = require('../../lib/languages')
 const pageMarker = chalk.green('|')
 const recordMarker = chalk.grey('.')
 const port = 4002
