+++ script/search/lunr-get-index-names.js
@@ -2,5 +2,5 @@ const fs = require('fs').promises
 const path = require('path')
 
 module.exports = async function getIndexNames () {
-  return await fs.readdir(path.join(__dirname, 'indexes'))
+  return await fs.readdir(path.join(__dirname, '../../lib/search/indexes'))
 }
