+++ tests/unit/search/parse-page-sections-into-records.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const path = require('path')
 const cheerio = require('cheerio')
-const parsePageSectionsIntoRecords = require('../../../lib/search/parse-page-sections-into-records')
+const parsePageSectionsIntoRecords = require('../../../script/search/parse-page-sections-into-records')
 const fixtures = {
   pageWithSections: fs.readFileSync(path.join(__dirname, 'fixtures/page-with-sections.html'), 'utf8'),
   pageWithoutSections: fs.readFileSync(path.join(__dirname, 'fixtures/page-without-sections.html'), 'utf8')
